Feature/refactory battle services

+++ app/services/adventure.php
@@ -1,21 +1,20 @@
 <?php
   header('Content-type: application/json');
   header('Access-Control-Allow-Origin: *');  
-  include_once("../repository/adventureDao.class.php");
-  include_once("../entity/Adventure.class.php");
+  include_once "/var/www/html/repository/adventureDao.class.php";
+  include_once "/var/www/html/entity/Adventure.class.php";
 
-  $msg = "";
   $result = "";
   $adventures = [];
   $adventureDao = new AdventureDao();
   $adventure = new Adventure();
   $json = file_get_contents('php://input');
   $post = json_decode($json);
-
+  $id = $_GET['id'];
   switch ($_SERVER['REQUEST_METHOD']) {
     case "GET":
-        if(isset($_GET['id'])){
-          $adventures = $adventureDao->findOne($_GET['id']);
+        if(isset($id) && $id] != null){
+          $adventures = $adventureDao->findOne($id);
           $result = array ('adventures'=>$adventures);
         }else{
             $adventures = $adventureDao->findAll();
@@ -27,9 +26,10 @@
         $adventureDao->insert($adventure);
         break;
     case "DELETE":
-        $id = $_GET['id'];
-        $msg =  $adventureDao->delete($id);
-        $result = array ('msg'=>$msg);
+        if(isset($id) && $id != null){
+          $msg =  $adventureDao->delete($id);
+          $result = array ('msg'=>$msg);
+        }
         break;
 }
