+++ app/services/world.php
@@ -1,7 +1,7 @@
 <?php
   header('Content-type: application/json');
   header('Access-Control-Allow-Origin: *');  
-  include_once("../repository/worldDao.class.php");
+  include_once "/var/www/html/repository/worldDao.class.php";
 
   $json = file_get_contents('php://input');
   $post = json_decode($json);
@@ -10,11 +10,11 @@
   $worlds = "";
   $world = "";
   $worldDao = new WorldDao();
-
+  $id = $_GET['id'];
   switch ($_SERVER['REQUEST_METHOD']) {
     case "GET":
-        if(isset($_GET['id'])){
-          $world = $worldDao->findOne($_GET['id']);
+        if(isset($id)){
+          $world = $worldDao->findOne($id);
           $result = array ('world'=>$world);
         }else{
             $worlds = $worldDao->findAll();
@@ -30,9 +30,10 @@
         $worldDao->insert($world);
         break;
     case "DELETE":
-        $id = $_GET['id'];
-        $msg =  $worldDao->delete($id);
-        $result = array ('msg'=>$msg);
+      if(isset($id) &&  $id != null){
+          $msg =  $worldDao->delete($id);
+          $result = array ('msg'=>$msg);
+        }
         break;
 }
